Django move TypeCollectivite to core.enums
