+++ django/docurba/core/admin.py
@@ -6,13 +6,13 @@ from typing import Any
 from django.contrib import admin
 from django.db import models
 
+from docurba.core.enums import TypeCollectivite
 from docurba.core.models import (
     Collectivite,
     Commune,
     Event,
     Procedure,
     Topic,
-    TypeCollectivite,
 )
