+++ django/docurba/core/enums.py
@@ -42,3 +42,21 @@ class ProcedureType(models.TextChoices):
             cls.REVISION_SIMPLIFIEE,
         ]
         return {member.value: member for member in cls if member in enums}
+
+
+class TypeCollectivite(models.TextChoices):
+    COM = "COM", "Commune"
+    COMD = "COMD", "Commune déléguée"
+    COMA = "COMA", "Commune associée"
+    CC = "CC", "Communauté de communes"
+    SMF = "SMF", "Syndicat Mixte Fermé"
+    SMO = "SMO", "Syndicat Mixte Ouvert"
+    METRO = "METRO", "Métropole"
+    CU = "CU", "Communauté Urbaine"
+    PETR = "PETR", "Pôle d'Équilibre Territorial et Rural"
+    MET69 = "MET69", "Métropole de Lyon"
+    SIVU = "SIVU", "Syndicat Intercommunal à Vocation Unique"
+    EPT = "EPT", "Établissement Public Territorial"
+    CA = "CA", "Communauté d'Agglomération"
+    POLEM = "POLEM", "Pôle Métropolitain"
+    SIVOM = "SIVOM", "Syndicat Intercommunal à Vocation Multiple"
