+++ django/docurba/core/models.py
@@ -15,29 +15,13 @@ from django.db.models.functions import Now
 from django.urls import reverse
 from django.utils import timezone
 
-from docurba.core.enums import CommuneType
+from docurba.core.enums import CommuneType, TypeCollectivite
 from docurba.core.utils import OversizedIndex
 
 logger = logging.getLogger(__name__)
 
 
 # NOTE(cms): These TextChoices should be moved to enums.py
-class TypeCollectivite(models.TextChoices):
-    COM = "COM", "Commune"
-    COMD = "COMD", "Commune déléguée"
-    COMA = "COMA", "Commune associée"
-    CC = "CC", "Communauté de communes"
-    SMF = "SMF", "Syndicat Mixte Fermé"
-    SMO = "SMO", "Syndicat Mixte Ouvert"
-    METRO = "METRO", "Métropole"
-    CU = "CU", "Communauté Urbaine"
-    PETR = "PETR", "Pôle d'Équilibre Territorial et Rural"
-    MET69 = "MET69", "Métropole de Lyon"
-    SIVU = "SIVU", "Syndicat Intercommunal à Vocation Unique"
-    EPT = "EPT", "Établissement Public Territorial"
-    CA = "CA", "Communauté d'Agglomération"
-    POLEM = "POLEM", "Pôle Métropolitain"
-    SIVOM = "SIVOM", "Syndicat Intercommunal à Vocation Multiple"
 
 
 class TypeDocument(models.TextChoices):
